Improve the heuristics of the bold broken check
Putting a span straight in body is not a good check because it will only
succeed for a monospaced body font, where the monospace font selection
is more likely to be placed on `.terminal`. For better realism still,
I’ve put it as a span inside a div inside a div.terminal, which matches
what the real terminal will do, so it should be as sound a check as can
be done short of doing it on the actual terminal (though quite frankly I
don’t see why we don’t just do that).